chore: begin breaking the vault logic into helper functions
