+++ packages/vault/src/credentials/README.md
@@ -1,0 +1,1 @@
+# Credential Store
