+++ packages/vault/src/objects/README.md
@@ -1,0 +1,1 @@
+# Objects Store
