+++ packages/vault/src/token-info/README.md
@@ -1,0 +1,1 @@
+# Token Info Store
