+++ packages/vault/src/vault/README.md
@@ -1,0 +1,1 @@
+# Vault Store
