+++ packages/vault/src/vault/utils/index.ts
@@ -1,0 +1,4 @@
+export * from './get-current-wallet'
+export * from './sync-account-info'
+export * from './sync-transactions'
+export * from './sync-wallet'
