+++ packages/vault/src/web-provider/README.md
@@ -1,0 +1,1 @@
+# Web Provider Store
